--- app/assets/javascripts/channels/conversation.coffee
@@ -1,12 +1,0 @@
-App.conversation = App.cable.subscriptions.create "ConversationChannel",
-  connected: ->
-    # Called when the subscription is ready for use on the server
-
-  disconnected: ->
-    # Called when the subscription has been terminated by the server
-
-  received: (data) ->
-    # Called when there's incoming data on the websocket for this channel
-
-  speak: ->
-    @perform 'speak'
