--- app/channels/application_cable/channels.rb
@@ -1,4 +1,0 @@
-module ApplicationCable
-  class Channels < ApplicationCable::Channel::Base
-  end
-end
