+++ app/channels/conversation_channel.rb
@@ -1,0 +1,3 @@
+# Be sure to restart your server when you modify this file. Action Cable runs in a loop that does not support auto reloading.
+class ConversationChannel < ApplicationCable::Channel
+end
