--- app/channels/conversations_channel.rb
@@ -1,19 +1,0 @@
-# Be sure to restart your server when you modify this file. Action Cable runs in a loop that does not support auto reloading.
-class ConversationsChannel < ApplicationCable::Channel
-  include SessionsHelper
-  def subscribed
-    stream_from "conversations-#{current_user.id}"
-  end
-
-  def unsubscribed
-    stop_all_streams
-  end
-
-  def speak(data)
-    message_params = data['message'].each_with_object({}) do |el, hash|
-      hash[el.values.first] = el.values.last
-    end
-
-    Message.create(message_params)
-    end
-end
