+++ spec/controllers/channels_controller.rb
@@ -1,0 +1,6 @@
+
+require 'rails_helper'
+
+RSpec.describe Channel, type: :controller do
+
+end
