--- spec/controllers/chats_controller_spec.rb
@@ -1,9 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe ChatsController, type: :controller do
-
-	before :each do
-		@person = FactoryGirl.create(:person)
-	end
-
-end
