--- spec/controllers/conversations_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe ConversationsController, type: :controller do
-
-end
