+++ spec/helpers/sessions_helper_spec.rb
@@ -11,9 +11,9 @@ require 'rails_helper'
 #   end
 # end
 RSpec.describe SessionsHelper, type: :helper do
-	#before :each do
-	#	@current_person = FactoryGirl.create(:person)
-	#end
+	before :each do
+		@current_person = FactoryGirl.create(:person)
+	end
 
 	#describe "set current coach" do
 	#	it "current_coach already assigned" do
@@ -21,24 +21,24 @@ RSpec.describe SessionsHelper, type: :helper do
 	#		expect(current_coach).to eq(@current_coach)
 	#	end
 
-#		it "assign a coach to current_coach" do
-#			expect(current_coach).to eq(nil)
-#			coach = FactoryGirl.create(:coach, person_id: current_person.id)
-#			expect(current_coach).to eql(coach)
-#
-#		end
-#	end
-#
-#	describe "set current user" do
-#		it "current_user already assigned" do
-#			@current_user = FactoryGirl.create(:user)
-#			expect(current_user).to eq(@current_user)
-#		end
-#
-#		it "assign a user to current_user" do
-#			expect(current_user).to eq(nil)
-#			user = FactoryGirl.create(:user, person_id: current_person.id)
-#			expect(current_user).to eql(user)
-#		end
-#	end
+	#	it "assign a coach to current_coach" do
+	#		expect(current_coach).to eq(nil)
+	#		coach = FactoryGirl.create(:coach, person_id: @current_person.id)
+	#		expect(current_coach).to eql(coach)
+
+	#	end
+	#end
+
+	#describe "set current user" do
+	#	it "current_user already assigned" do
+	#		@current_user = FactoryGirl.create(:user)
+	#		expect(current_user).to eq(@current_user)
+	#	end
+
+	#	it "assign a user to current_user" do
+	#		expect(current_user).to eq(nil)
+	#		user = FactoryGirl.create(:user, person_id: @current_person.id)
+	#		expect(current_user).to eql(user)
+	#	end
+	#end
 end
