--- spec/jobs/message_broadcast_job.rb
@@ -1,4 +1,0 @@
-require "rails_helper"
-
-RSpec.describe MessageBroadcastJob, :type => :job do
-end
