--- spec/mailers/application_mailer_spec.rb
@@ -1,2 +1,0 @@
-RSpec.describe ApplicationMailer, type: :mailer do
-end
