--- spec/mailers/user_mailer_spec.rb
@@ -1,2 +1,0 @@
-RSpec.describe UserMailer, type: :mailer do
-end
